Move async_migrate_entry() to __init__.py.
